+++ args.py
@@ -50,7 +50,7 @@ def parse_args(mode='train'):
     parser.add_argument('--name', default='model', type=str, help='file & run name')
     
     ### config ####
-    parser.add_argument('--config_path', default='./config/config.yml', type=str, help='file & run name')
+    parser.add_argument('--config_path', default='./config_dir/config.yml', type=str, help='file & run name')
     parser.add_argument('--config', default='lstm', type=str, help='select config')
     
     args = parser.parse_args()
